requirements.txt
numpy
matplotlibsklearn
sentencepiece
tensorflow >= 2